chore: migrate from `psalm` to `mago` (#527)

+++ src/Psl/Str/ord.php
@@ -21,5 +21,5 @@ use function mb_ord;
  */
 function ord(string $character, Encoding $encoding = Encoding::Utf8): int
 {
-    return (int) mb_ord($character, $encoding->value);
+    return mb_ord($character, $encoding->value);
 }
